add Windows tests for executor/rawexec
